requirements.txt
# Core ML/CV Libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Geospatial
geopy>=2.3.0
gpxpy>=1.5.0
pyproj>=3.5.0

# API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
jinja2>=3.1.0

# Data Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0

# Evaluation & Metrics
scikit-learn>=1.3.0
scipy>=1.11.0

# Video Processing
moviepy>=1.0.3

# Deployment
gunicorn>=21.2.0

# Development & Testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0
